chore: split e2e_admin_1 test suite
